+++ text2qti/cmdline.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (c) 2020, Geoffrey M. Poore
+# Copyright (c) 2020-2021, Geoffrey M. Poore
 # All rights reserved.
 #
 # Licensed under the BSD 3-Clause License:
@@ -11,12 +11,16 @@
 import argparse
 import os
 import pathlib
+import platform
+import shutil
+import subprocess
 import textwrap
 from .version import __version__ as version
 from .err import Text2qtiError
 from .config import Config
 from .quiz import Quiz
 from .qti import QTI
+from .export import quiz_to_pandoc_markdown
 
 
 
@@ -34,6 +38,13 @@ def main():
                         help='Allow special code blocks to be executed and insert their output (off by default for security)')
     parser.add_argument('--pandoc-mathml', action='store_const', const=True,
                         help='Convert LaTeX math to MathML using Pandoc (this will create a cache file "_text2qti_cache.zip" in the quiz file directory)')
+    soln_group = parser.add_mutually_exclusive_group()
+    soln_group.add_argument('--solutions', nargs=1, metavar='SOLUTIONS_FILE',
+                            help='Save solutions in Pandoc Markdown (.md) or PDF (.pdf) format, and create a QTI file '
+                                 '(Pandoc Markdown output is only suitable for use with LaTeX; PDF output requires Pandoc plus LaTeX)')
+    soln_group.add_argument('--only-solutions', nargs=1, metavar='SOLUTIONS_FILE',
+                            help='Save solutions in Pandoc Markdown (.md) or PDF (.pdf) format, and do not create a QTI file '
+                                 '(Pandoc Markdown output is only suitable for use with LaTeX; PDF output requires Pandoc plus LaTeX)')
     parser.add_argument('file',
                         help='File to convert from text to QTI')
     args = parser.parse_args()
@@ -86,10 +97,41 @@ def main():
         raise Text2qtiError(f'File "{file_path}" is not encoded in valid UTF-8:\n{e}')
 
     cwd = pathlib.Path.cwd()
+    if args.solutions:
+        qti_path = pathlib.Path(f'{file_path.stem}.zip')
+        solutions_path = pathlib.Path(args.solutions[0]).expanduser().absolute()
+    elif args.only_solutions:
+        qti_path = None
+        solutions_path = pathlib.Path(args.only_solutions[0]).expanduser().absolute()
+    else:
+        qti_path = pathlib.Path(f'{file_path.stem}.zip')
+        solutions_path = None
     os.chdir(file_path.parent)
     try:
         quiz = Quiz(text, config=config, source_name=file_path.as_posix())
-        qti = QTI(quiz)
-        qti.save(f'{file_path.stem}.zip')
+        if solutions_path is not None:
+            if solutions_path.suffix.lower() in ('.md', '.markdown'):
+                solutions_path.write_text(quiz_to_pandoc_markdown(quiz, solutions=True), encoding='utf8')
+            elif solutions_path.suffix.lower() == '.pdf':
+                if not shutil.which('pandoc'):
+                    raise Text2qtiError('Exporting solutions in PDF format requires Pandoc (https://pandoc.org/)')
+                if not shutil.which('pdflatex'):
+                    raise Text2qtiError('Exporting solutions in PDF format requires LaTeX (https://www.tug.org/texlive/ or https://miktex.org/)')
+                if platform.system() == 'Windows':
+                    cmd = [shutil.which('pandoc'), '-f', 'markdown', '-o', str(solutions_path)]
+                else:
+                    cmd = ['pandoc', '-f', 'markdown', '-o', str(solutions_path)]
+                subprocess.run(
+                    cmd,
+                    input=quiz_to_pandoc_markdown(quiz, solutions=True),
+                    capture_output=True,
+                    check=True,
+                    encoding='utf8'
+                )
+            else:
+                raise Text2qtiError(f'Unsupported export format {solutions_path.suffix} for solutions; use .md, .markdown, or .pdf')
+        if qti_path is not None:
+            qti = QTI(quiz)
+            qti.save(qti_path)
     finally:
         os.chdir(cwd)